% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_kapferer.R
\name{create_kapferer}
\alias{create_kapferer}
\title{Generate Kapferer results}
\usage{
create_kapferer(files)
}
\arguments{
\item{files}{The path to the folder containing all the conflict files}
}
\description{
This is a convenience function that does all the kapferer calcs and outputs useful results
This function DOES NOT WORK at the moment. I am doubtful that this dataset is good enugh quality to actually use
}
