% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_reweighter.R
\name{edge_reweighter}
\alias{edge_reweighter}
\title{Re wieghter

This function re-weights the values of the edges to explore different edge weigh ratios}
\usage{
edge_reweighter(g, weight, min_k, mid_point, range_k, reverse_order = FALSE)
}
\arguments{
\item{g}{An igraph object}

\item{weight}{the edge weight attribute of the graph}

\item{min_k}{a positive numeric. The minimum value of K}

\item{mid_point}{an integer the midpoint between the minimum and maximum values}

\item{range_k}{a positive numeric. The maximum multilplier between the smallest and largest k values}

\item{reverse_order}{logical. high weights become low weight and visa versa}
}
\description{
Re wieghter

This function re-weights the values of the edges to explore different edge weigh ratios
}
