% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dl_graph.R
\name{load_dl_graph}
\alias{load_dl_graph}
\title{load dl graph type}
\usage{
load_dl_graph(file_path, directed = FALSE, return_graph = TRUE)
}
\arguments{
\item{file_path}{A character string. The file path of the dl file to be loaded}

\item{directed}{A logical. whether the graph is directed or not}

\item{return_graph}{Logical. whether a graph or adjacency matrix should be returned.}
}
\value{
The function returns a named list where each element of the list is a an igraph object representing each network in the dl file
}
\description{
a dl loader for pajek dl files and dl files from the Linton freedman network archive http://moreno.ss.uci.edu
}
\details{
According to the igraph documentation there is a lack of dl format standard. This means that creating a dl loader is challenging.
unfortunately a lot of file on the pajek database are in dl format. I have made this loader so that they can be used.
}
