% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medici_function.R
\name{medici_function}
\alias{medici_function}
\title{Medici function}
\usage{
medici_function(medici = "MEDICI", oligarch = "STROZZI")
}
\arguments{
\item{medici}{a character string. The name of the Medici party who will act as the beligerent. must be all caps}

\item{oligarch}{a character string. The name of the Oligarch party who will act as the beligerent. must be all caps}
}
\description{
This function is a horror against all that is good in programming... I care not.
}
\details{
This function produces two dataframes for the florentine families. It is big and ugly but a better solution doesn't seem to be valuable
}
