% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medici_process.R
\name{medici_process}
\alias{medici_process}
\title{The medici code
This is a bit of a Frankenstien that allows the medici analysis to be run in relateivley intelligable chunks
The data is loaded several times, becuase I can't be bothered to finness the process}
\usage{
medici_process(edge_file, node_file)
}
\arguments{
\item{edge_file}{the path to file containing edge data}

\item{node_file}{the path to the file containing node data}
}
\description{
The medici code
This is a bit of a Frankenstien that allows the medici analysis to be run in relateivley intelligable chunks
The data is loaded several times, becuase I can't be bothered to finness the process
}
