% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_clusters.R
\name{rename_clusters}
\alias{rename_clusters}
\title{Rename_clusters}
\usage{
rename_clusters(df, vars, ref_var = "node", priority = "medici")
}
\arguments{
\item{df}{A dataframe. The data frame with the clusters and the node names in}

\item{vars}{A character vector. A vector of characters naming the columns that need to be matched with beligerents}
}
\description{
used for compareing SETSe vs other clustering methods against a ground truth. Very niche.
}
