% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thurman_office_function.R
\name{thurman_office_function}
\alias{thurman_office_function}
\title{The Thurman office function}
\usage{
thurman_office_function(belligerent_a, belligerent_b, remove_minna = TRUE)
}
\arguments{
\item{belligerent_a}{character string. The first person in the conflict}

\item{belligerent_b}{character string/ The second person in the conflict}
}
\value{
A list of three elements
\itemize{
\item The full node embedding for each combination of edge strength
\item The summary showing the aggregated factional control of the group
\item The embeddings in igraph format
}
}
\description{
The sugar to make process the thurman conflicts easier. Beligerent A is taken as the baseline
}
