% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_beligerent.R
\name{two_beligerent_conflict}
\alias{two_beligerent_conflict}
\title{two_beligerent_conflict}
\usage{
two_beligerent_conflict(g, beligerents)
}
\arguments{
\item{g}{an Igraph object}

\item{beligerents}{numeric vector of length two. Node indices of the the beligerent nodes}
}
\description{
creates the forces for a conflict involving only two beligerents. The first node is always positive
}
